# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationExpectation": {
        "properties": {
          "check_type": {
            "title": "Check Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expected_result": {
            "title": "Expected Result",
            "type": "string"
          },
          "fix_command_execution_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fix Command Execution Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "relation_to_customer_symptom": {
            "title": "Relation To Customer Symptom",
            "type": "string"
          },
          "required": {
            "default": true,
            "title": "Required",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "status": {
            "default": "pending",
            "title": "Status",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "validation_result_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Result Id"
          }
        },
        "required": [
          "id",
          "run_id",
          "check_type",
          "expected_result",
          "relation_to_customer_symptom",
          "created_at"
        ],
        "title": "ValidationExpectation",
        "type": "object"
      },
      "ValidationResult": {
        "properties": {
          "check_type": {
            "title": "Check Type",
            "type": "string"
          },
          "command_execution_id": {
            "title": "Command Execution Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence": {
            "title": "Evidence",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "required": [
          "id",
          "run_id",
          "command_execution_id",
          "check_type",
          "passed",
          "summary",
          "evidence",
          "created_at"
        ],
        "title": "ValidationResult",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Backend control plane for Phoenix ERP, technician approvals, SSH execution, and audit state.",
    "title": "techbold AI Service Desk Autopilot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs/{run_id}/validation-expectations": {
      "get": {
        "operationId": "get_run_validation_expectations_api_runs__run_id__validation_expectations_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ValidationExpectation"
                  },
                  "title": "Response Get Run Validation Expectations Api Runs  Run Id  Validation Expectations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Validation Expectations"
      }
    },
    "/api/runs/{run_id}/validation-results": {
      "get": {
        "operationId": "get_run_validation_results_api_runs__run_id__validation_results_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ValidationResult"
                  },
                  "title": "Response Get Run Validation Results Api Runs  Run Id  Validation Results Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Validation Results"
      }
    },
...
  }
}